Add a universal Wine path patch with env vars
Instead of relying on two seperate builds for proot/chroot and glibc, use only one and utilize env vars to change hardcoded path locations for Wine. What this will allow is to streamline builds and reduce the amount of builds I will need to do. Since all of our environments utilize glibc and Wine built using this method is compatible across the board, this should be a "one-stop-shop" solution for all of our problems. 

This is still a very crude implementation and it might be possible to do a dynamic one which will detect an environment you are using and automatically link against appropriate directories/libraries.

P.S. This is a very experimental patch which might not work in our favor.

Usage for this patch is like this (these are crucial for Wine to be functional):

```export WINE_CA_CERT_PATH=path/to/cert/location
export WINE_TMP_PATH=path/to/tmp/dir
export WINE_SHARE_DIRS=path/to/share/dir```